+++ crates/common/Cargo.toml
@@ -6,11 +6,14 @@ edition = "2021"
 [lib]
 
 [dependencies]
+dcl_component = { workspace = true }
+
 bevy = { workspace = true }
 ethers-core = { workspace = true }
 futures-lite = { workspace = true }
 tokio = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
 
 hex = "0.4.3"
 smallvec = "1.11"
