+++ crates/common/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod dynamics;
+pub mod profile;
 pub mod rpc;
 pub mod sets;
 pub mod structs;
