+++ crates/dcl/src/js/mod.rs
@@ -28,6 +28,7 @@ pub mod fetch;
 pub mod portables;
 pub mod restricted_actions;
 pub mod runtime;
+pub mod user_identity;
 
 // marker to indicate shutdown has been triggered
 pub struct ShuttingDown;
@@ -71,12 +72,13 @@ pub fn create_runtime(init: bool) -> JsRuntime {
 
     let mut ops = vec![op_require::DECL, op_log::DECL, op_error::DECL];
 
-    let op_sets: [Vec<deno_core::OpDecl>; 5] = [
+    let op_sets: [Vec<deno_core::OpDecl>; 6] = [
         engine::ops(),
         restricted_actions::ops(),
         runtime::ops(),
         fetch::ops(),
         portables::ops(),
+        user_identity::ops(),
     ];
 
     // add plugin registrations
