+++ crates/dcl/src/js/modules/UserIdentity.js
@@ -1,2 +1,6 @@
 module.exports.getUserPublicKey = async function (body) { return { address: undefined } }
-module.exports.getUserData = async function (body) { return {} }
+module.exports.getUserData = async function (body) { 
+    return {
+        data: Deno.core.ops.op_get_user_data()
+    };
+}
