+++ crates/restricted_actions/Cargo.toml
@@ -11,6 +11,7 @@ ipfs = { workspace = true }
 scene_runner = { workspace = true }
 avatar = { workspace = true }
 ui_core = { workspace = true }
+comms = { workspace = true }
 
 bevy = { workspace = true }
 tokio = { workspace = true }
